+++ admin/tabs/clients/activity/ClientDetailModel.ts
@@ -83,11 +83,11 @@ export class ClientDetailModel extends HoistModel implements ActivityDetailProvi
                         filters: {field: 'tabId', op: '=', value: tabId}
                     }
                 });
-            })
-            .catch(e => {
-                if (loadSpec.isStale || loadSpec.isAutoRefresh) return;
-                XH.handleException(e, {alertType: 'toast'});
-                this.trackLogs = [];
             });
+    }
+
+    override handleLoadException(e: unknown) {
+        XH.handleException(e, {alertType: 'toast'});
+        this.trackLogs = [];
     }
 }