+++ admin/tabs/cluster/instances/connpool/ConnPoolMonitorModel.ts
@@ -133,8 +133,7 @@ export class ConnPoolMonitorModel extends BaseInstanceModel {
                         step: true
                     }
                 ]);
-            })
-            .catch(e => this.handleLoadException(e, loadSpec));
+            });
     }
 
     async takeSnapshotAsync() {