+++ core/model/HoistModel.ts
@@ -17,7 +17,7 @@ import {
     TaskObserver
 } from '../';
 import {instanceManager} from '../impl/InstanceManager';
-import {Loadable, LoadSpec, LoadSupport} from '../load';
+import {AbortMode, Loadable, LoadSpec, LoadSupport} from '../load';
 import {ModelSelector} from './';
 import {Class} from 'type-fest';
 
@@ -198,7 +198,13 @@ export abstract class HoistModel extends HoistBase implements Loadable {
         return this.loadSupport?.autoRefreshAsync(meta);
     }
 
+    abortMode: AbortMode = 'onStale';
+
     async doLoadAsync(loadSpec: LoadSpec) {}
+
+    handleLoadException(e: unknown, loadSpec: LoadSpec): void | Promise<void> {
+        return this.loadSupport?.handleLoadException(e, loadSpec);
+    }
 
     async loadAsync(loadSpec?: LoadSpecConfig | CallContext) {
         return this.loadSupport?.loadAsync(loadSpec);